Funcionalidad de gestión y de agendamiento funcionales

+++ cl.acaya.cobranza.business/cl.acaya.business.dao-ejb/src/main/java/cl/acaya/cobranza/business/daoEjb/dao/ContingenciaDAO.java
@@ -5,7 +5,7 @@ import cl.acaya.cobranza.business.daoEjb.entities.Contingencia;
 import java.util.List;
 
 
-public interface ContingenciaDAO extends GenericDAO<Contingencia, Long> {
+public interface ContingenciaDAO extends GenericDAO<Contingencia, Integer> {
 
     List<Contingencia> findAll();
 }
