+++ cl.acaya.cobranza.business/cl.acaya.business.dao-ejb/src/main/java/cl/acaya/cobranza/business/daoEjb/dao/impl/ContingenciaDAOimpl.java
@@ -12,7 +12,7 @@ import java.util.List;
  */
 @Local(ContingenciaDAO.class)
 @Stateless
-public class ContingenciaDAOimpl extends GenericDAOImpl<Contingencia,Long> implements ContingenciaDAO {
+public class ContingenciaDAOimpl extends GenericDAOImpl<Contingencia,Integer> implements ContingenciaDAO {
 
     public ContingenciaDAOimpl() {
         super(Contingencia.class);
